+++ changelog.md
@@ -5,6 +5,7 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add a new `copy()` method to `Group`.
 - Remove `CURRENT_OS_ID` and `CURRENT_OS_LABEL`.
 
 ## [1.5.0 (2024-11-10)](https://github.com/kdeldycke/extra-platforms/compare/v1.4.0...v1.5.0)